agregada tarea 2

+++ Unidad-1/Tareas/tarea2/tarea2/src/main/java/Ejercicio8.java
@@ -18,7 +18,7 @@ public class Ejercicio8 {
 
         do{
             System.out.println("Dame la nota:");
-            nota=sc.nextInt();
+            nota=sc.nextDouble();
 
             cuenta=cuenta+1;
             acumulador=acumulador+nota;
